Bugfix:   * Do not allow topics with names beginning with "Special:".
git-svn-id: https://jamwiki.svn.sourceforge.net/svnroot/jamwiki/wiki/trunk@746 f65b7e76-0091-dd46-87ae-4bc0c949c03e

+++ src/java/org/jamwiki/utils/Utilities.java
@@ -464,6 +464,7 @@ public class Utilities {
 	 */
 	public static boolean validateName(String name) {
 		if (!StringUtils.hasText(name)) return false;
+		if (name.toLowerCase().startsWith(WikiBase.NAMESPACE_SPECIAL.toLowerCase())) return false;
 		// try to remove invalid characters
 		String cleaned = StringUtils.deleteAny(name, "\n\r\"\'><{}#/\\=[]");
 		return name.equals(cleaned);
